# specs/openapi.yaml
openapi: 3.1.0
info:
  title: YouTube Curator API
  version: 0.2.0
  description: API privada para la aplicación personal YouTube Curator.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /channels:
    get:
      tags: [Channels]
      summary: Listar canales
      operationId: listChannels
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: unclassified
          in: query
          schema:
            type: boolean
        - name: subscribed
          in: query
          schema:
            type: boolean
        - name: query
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Canales paginados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPage"
  /channels/{channelId}:
    get:
      tags: [Channels]
      summary: Obtener canal
      operationId: getChannel
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      responses:
        "200":
          description: Canal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          $ref: "#/components/responses/NotFound"
  /channels/{channelId}/block:
    put:
      tags: [Channels]
      summary: Bloquear o desbloquear canal
      operationId: setChannelBlocked
      parameters:
        - $ref: "#/components/parameters/ChannelId"
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blocked]
              properties:
                blocked:
                  type: boolean
      responses:
        "200":
          description: Estado de bloqueo actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: integer
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
  responses:
...
    NotFound:
      description: Recurso inexistente.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Channel:
      type: object
      required:
        - id
        - youtubeChannelId
        - title
        - subscribed
        - locallyFollowed
        - blocked
        - categoryIds
      properties:
        id:
          type: integer
        youtubeChannelId:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: [string, "null"]
          format: uri
        subscribed:
          type: boolean
        locallyFollowed:
          type: boolean
        blocked:
          type: boolean
        categoryIds:
          type: array
          items:
            type: integer
    ChannelPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Channel"
        nextCursor:
          type: [string, "null"]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true